Got reminder thread running in background
